requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
flask_restx